Refactor: interface 이름 규칙 적용 & 파일 분리
