+++ src/hooks/queries/useRandomMv.tsx
@@ -1,12 +1,12 @@
 import { useMutation } from 'react-query';
 import { privateInstance } from '../../library/apis/axiosInstance';
-interface RandomMvData {
+interface IRandomMv {
   selectId: string;
   page: number;
 }
 
 export const useRandomMv = () => {
-  const getRandomMv = async (data: RandomMvData) => {
+  const getRandomMv = async (data: IRandomMv) => {
     const { selectId, page } = data;
     // console.log(selectId);
     const response = await privateInstance.post(
