+++ src/types/RandomMv.ts
@@ -5,12 +5,6 @@ export interface IVideoData {
   singer: string;
 }
 
-export interface IVideoInfoProps {
-  title: string;
-  views: string;
-  onAddButtonClick?: () => void;
-}
-
 export interface IVideoUrl {
   url: string;
 }
