+++ src/types/user.ts
@@ -12,9 +12,3 @@ export interface User {
     refresh: string;
   };
 }
-
-export interface ISignup {
-  email: string;
-  password?: string;
-  type: string;
-}
